+++ bee-api/bee-rest-api/src/endpoints/mod.rs
@@ -39,7 +39,7 @@ pub(crate) type Bech32Hrp = String;
 
 pub(crate) const CONFIRMED_THRESHOLD: u32 = 5;
 
-pub async fn init_full_node<N: Node>(
+pub fn init_full_node<N: Node>(
     rest_api_config: RestApiConfig,
     protocol_config: ProtocolConfig,
     network_id: NetworkId,
@@ -109,7 +109,7 @@ where
                 requested_messages,
                 consensus_worker,
             )
-            .recover(handle_rejection);
+            .recover(|err| async { handle_rejection(err) });
 
             let (_, server) =
                 warp::serve(routes).bind_with_graceful_shutdown(rest_api_config.bind_socket_addr(), async {
@@ -125,7 +125,7 @@ where
     }
 }
 
-async fn handle_rejection(err: Rejection) -> Result<impl Reply, Infallible> {
+fn handle_rejection(err: Rejection) -> Result<impl Reply, Infallible> {
     let (http_code, err_code, reason) = match err.find() {
         // handle custom rejections
         Some(CustomRejection::Forbidden) => (StatusCode::FORBIDDEN, "403", "access forbidden"),
@@ -153,7 +153,7 @@ async fn handle_rejection(err: Rejection) -> Result<impl Reply, Infallible> {
     ))
 }
 
-pub async fn init_entry_node<N: Node>(rest_api_config: RestApiConfig, node_builder: N::Builder) -> N::Builder
+pub fn init_entry_node<N: Node>(rest_api_config: RestApiConfig, node_builder: N::Builder) -> N::Builder
 where
     N::Backend: StorageBackend,
 {
@@ -174,7 +174,9 @@ where
         node.spawn::<Self, _, _>(|shutdown| async move {
             info!("Running.");
 
-            let health = warp::path("health").map(|| StatusCode::OK).recover(handle_rejection);
+            let health = warp::path("health")
+                .map(|| StatusCode::OK)
+                .recover(|err| async { handle_rejection(err) });
 
             let (_, server) = warp::serve(health).bind_with_graceful_shutdown(config.bind_socket_addr(), async {
                 shutdown.await.ok();