+++ bee-protocol/src/workers/message/payload/milestone.rs
@@ -19,7 +19,7 @@ use bee_message::{
     Message, MessageId,
 };
 use bee_runtime::{event::Bus, node::Node, shutdown_stream::ShutdownStream, worker::Worker};
-use bee_tangle::{event::LatestMilestoneChanged, MessageRef, Tangle, TangleWorker};
+use bee_tangle::{event::LatestMilestoneChanged, Tangle, TangleWorker};
 
 use async_trait::async_trait;
 use futures::{future::FutureExt, stream::StreamExt};
@@ -37,7 +37,7 @@ pub(crate) enum Error {
 
 pub(crate) struct MilestonePayloadWorkerEvent {
     pub(crate) message_id: MessageId,
-    pub(crate) message: MessageRef,
+    pub(crate) message: Message,
 }
 
 pub(crate) struct MilestonePayloadWorker {
@@ -68,10 +68,10 @@ fn validate(
 }
 
 #[allow(clippy::too_many_arguments)]
-async fn process<B: StorageBackend>(
+fn process<B: StorageBackend>(
     tangle: &Tangle<B>,
     message_id: MessageId,
-    message: MessageRef,
+    message: Message,
     peer_manager: &PeerManager,
     metrics: &NodeMetrics,
     requested_milestones: &RequestedMilestones,
@@ -90,7 +90,7 @@ async fn process<B: StorageBackend>(
 
         match validate(message_id, &message, milestone, key_manager) {
             Ok(milestone) => {
-                tangle.add_milestone(index, milestone.clone()).await;
+                tangle.add_milestone(index, milestone.clone());
                 if index > tangle.get_latest_milestone_index() {
                     info!("New milestone {} {}.", index, milestone.message_id());
                     tangle.update_latest_milestone_index(index);
@@ -166,8 +166,7 @@ where
                     &milestone_solidifier,
                     &key_manager,
                     &bus,
-                )
-                .await;
+                );
             }
 
             // Before the worker completely stops, the receiver needs to be drained for milestone payloads to be
@@ -187,8 +186,7 @@ where
                     &milestone_solidifier,
                     &key_manager,
                     &bus,
-                )
-                .await;
+                );
                 count += 1;
             }