+++ bee-protocol/src/workers/message/payload/mod.rs
@@ -11,9 +11,8 @@ pub(crate) use transaction::{TransactionPayloadWorker, TransactionPayloadWorkerE
 
 use crate::workers::storage::StorageBackend;
 
-use bee_message::{payload::Payload, MessageId};
+use bee_message::{payload::Payload, Message, MessageId};
 use bee_runtime::{node::Node, shutdown_stream::ShutdownStream, worker::Worker};
-use bee_tangle::MessageRef;
 
 use async_trait::async_trait;
 use futures::{future::FutureExt, stream::StreamExt};
@@ -25,16 +24,16 @@ use std::{any::TypeId, convert::Infallible};
 
 pub(crate) struct PayloadWorkerEvent {
     pub(crate) message_id: MessageId,
-    pub(crate) message: MessageRef,
+    pub(crate) message: Message,
 }
 
 pub(crate) struct PayloadWorker {
     pub(crate) tx: mpsc::UnboundedSender<PayloadWorkerEvent>,
 }
 
-async fn process(
+fn process(
     message_id: MessageId,
-    message: MessageRef,
+    message: Message,
     transaction_payload_worker: &mpsc::UnboundedSender<TransactionPayloadWorkerEvent>,
     milestone_payload_worker: &mpsc::UnboundedSender<MilestonePayloadWorkerEvent>,
     indexation_payload_worker: &mpsc::UnboundedSender<IndexationPayloadWorkerEvent>,
@@ -104,8 +103,7 @@ where
                     &transaction_payload_worker,
                     &milestone_payload_worker,
                     &indexation_payload_worker,
-                )
-                .await;
+                );
             }
 
             // Before the worker completely stops, the receiver needs to be drained for payloads to be analysed.
@@ -121,8 +119,7 @@ where
                     &transaction_payload_worker,
                     &milestone_payload_worker,
                     &indexation_payload_worker,
-                )
-                .await;
+                );
                 count += 1;
             }