+++ bee-storage/bee-storage-rocksdb/src/storage.rs
@@ -20,8 +20,8 @@ use bee_message::{
 
 use parking_lot::RwLock;
 use rocksdb::{
-    ColumnFamily, ColumnFamilyDescriptor, DBCompactionStyle, DBCompressionType, Env, FlushOptions, Options,
-    SliceTransform, DB,
+    ColumnFamily, ColumnFamilyDescriptor, DBCompactionStyle, DBCompressionType, Env, FlushOptions, MergeOperands,
+    Options, SliceTransform, DB,
 };
 
 pub(crate) const STORAGE_VERSION: StorageVersion = StorageVersion(9);
@@ -42,7 +42,14 @@ impl Storage {
 
         let cf_message_id_to_message = ColumnFamilyDescriptor::new(CF_MESSAGE_ID_TO_MESSAGE, Options::default());
 
-        let cf_message_id_to_metadata = ColumnFamilyDescriptor::new(CF_MESSAGE_ID_TO_METADATA, Options::default());
+        fn keep_current(_key: &[u8], existing_val: Option<&[u8]>, operands: &MergeOperands) -> Option<Vec<u8>> {
+            // Keep the existing value, if the value does not exist, take the first operand
+            // instead.
+            existing_val.or_else(|| operands.into_iter().next()).map(|v| v.to_vec())
+        }
+        let mut options = Options::default();
+        options.set_merge_operator_associative("keep current", keep_current);
+        let cf_message_id_to_metadata = ColumnFamilyDescriptor::new(CF_MESSAGE_ID_TO_METADATA, options);
 
         let mut options = Options::default();
         options.set_prefix_extractor(SliceTransform::create_fixed_prefix(MESSAGE_ID_LENGTH));