+++ bee-tangle/src/lib.rs
@@ -31,9 +31,6 @@ pub mod unreferenced_message;
 pub mod urts;
 
 mod conflict;
-mod vec_set;
-mod vertex;
-mod vertices;
 
 pub use conflict::ConflictReason;
 pub use tangle::Tangle;
@@ -41,24 +38,7 @@ pub use tangle_worker::TangleWorker;
 
 use tip_pool_cleaner_worker::TipPoolCleanerWorker;
 
-use crate::vec_set::VecSet;
-
-use bee_message::Message;
 use bee_runtime::node::{Node, NodeBuilder};
-
-use std::{ops::Deref, sync::Arc};
-
-/// A thread-safe reference to a `Message`.
-#[derive(Clone)]
-pub struct MessageRef(pub(crate) Arc<Message>);
-
-impl Deref for MessageRef {
-    type Target = Message;
-
-    fn deref(&self) -> &Self::Target {
-        &*self.0
-    }
-}
 
 /// Initiate the tangle on top of the given node builder.
 pub fn init<N: Node>(tangle_config: &config::TangleConfig, node_builder: N::Builder) -> N::Builder