+++ bee-storage/bee-storage-null/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.3.0 - 2022-03-17
+
+### Added
+
+- Implementation of `InsertStrict` for `Storage`;
+
 ## 0.2.0 - 2022-03-11
 
 ### Added