+++ bee-storage/bee-storage/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.11.0 - 2022-03-17
+
+### Added
+
+- `InsertStrict` access trait;
+
 ## 0.10.0 - 2022-03-11
 
 ### Added